% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-diagnostics.R
\name{plotPCA,BatchVariaData-method}
\alias{plotPCA,BatchVariaData-method}
\title{Plot PCA of expression data}
\usage{
\S4method{plotPCA}{BatchVariaData}(object, assays = "raw", colourBy = "batch", ...)
}
\arguments{
\item{object}{A \code{BatchVariaData} object.}

\item{assays}{assays to use for PCA (default \code{"raw"}).}

\item{colourBy}{Optional column name from \code{colData(object)} used to
colour samples.}

\item{...}{Further arguments, required by the generic; unused.}
}
\value{
A list of  \code{ggplot2} objects (one per assay).
}
\description{
Visualise global sample structure using Principal Component Analysis (PCA).
This function extracts an expression assay from a \code{BatchVariaData}
object, performs PCA on the sample-wise expression matrix, and plots the
first two principal components. Samples can optionally be coloured by a
variable in \code{colData}.
}
\details{
PCA is computed on the \strong{transposed expression matrix} so that samples
represent observations and genes represent variables. The first two
principal components are plotted.

If \code{colourBy} is specified, the corresponding variable from
\code{colData(object)} is used to colour points in the plot.
}
\examples{
set.seed(1)
bv <- exampleBatchVaria(nGenes = 100)

plotPCA(
    bv,
    assays = "raw",
    colourBy = "batch"
)

}
