% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchVariaData-methods.R
\name{show,BatchVariaData-method}
\alias{show,BatchVariaData-method}
\title{Show method for BatchVariaData}
\usage{
\S4method{show}{BatchVariaData}(object)
}
\arguments{
\item{object}{A BatchVariaData object}
}
\value{
prints a summary of \code{object}. Returns \code{object} invisibly.
}
\description{
Provides a concise summary of the container including assays,
sample annotations, and analysis history.
}
\examples{
set.seed(1)
bv <- exampleBatchVaria(nGenes = 100)
bv

}
