% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correction-registry.R
\name{unregisterCorrectionMethod}
\alias{unregisterCorrectionMethod}
\title{Remove a registered batch correction method}
\usage{
unregisterCorrectionMethod(name)
}
\arguments{
\item{name}{Character. Name of the method to remove.}
}
\value{
The removed name, invisibly.
}
\description{
Remove a registered batch correction method
}
\examples{
dummy <- function(assayMatrix, batch, sampleData, preserve = NULL, ...) {
    assayMatrix
}
registerCorrectionMethod("dummy", dummy)
unregisterCorrectionMethod("dummy")
availableCorrectionMethods()

}
\seealso{
\code{\link{registerCorrectionMethod}}
}
