% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance-engine-registry.R
\name{unregisterVarianceEngine}
\alias{unregisterVarianceEngine}
\title{Remove a registered variance profiling engine}
\usage{
unregisterVarianceEngine(name)
}
\arguments{
\item{name}{Character. Name of the engine to remove.}
}
\value{
The removed name, invisibly.
}
\description{
Remove a registered variance profiling engine
}
\examples{
dummy <- function(assayMatrix, formula, sampleData, ...) {
    newVarianceSummary(
        source = "dummy", term = "all",
        varianceFraction = 1, nFeatures = nrow(assayMatrix)
    )
}
registerVarianceEngine("dummy", dummy)
unregisterVarianceEngine("dummy")
availableVarianceMethods()

}
\seealso{
\code{\link{registerVarianceEngine}}
}
