% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchVariaData-accessors.R
\name{varianceHistory}
\alias{varianceHistory}
\title{Access variance profiling history}
\usage{
varianceHistory(object)
}
\arguments{
\item{object}{BatchVariaData object}
}
\value{
list of variance analysis blocks
}
\description{
Access variance profiling history
}
\examples{
set.seed(1)
bv <- exampleBatchVaria(nGenes = 100)
bv <- profileVariance(bv, ~batch)
length(varianceHistory(bv))

}
