% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correction-engines.R
\name{availableCorrectionMethods}
\alias{availableCorrectionMethods}
\title{Available correction methods}
\usage{
availableCorrectionMethods()
}
\value{
Character vector of method identifiers
}
\description{
Returns the names of supported correction engines.
}
\examples{
availableCorrectionMethods()
}
