% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-diagnostics.R
\name{plotBatchEntropy}
\alias{plotBatchEntropy}
\title{Plot batch mixing entropy}
\usage{
plotBatchEntropy(object, assayName = "raw", batchVar = "batch", k = 10)
}
\arguments{
\item{object}{BatchVariaData object}

\item{assayName}{Assay name}

\item{batchVar}{Column in colData representing batch}

\item{k}{Number of neighbours}
}
\value{
ggplot object
}
\description{
Computes local neighbourhood entropy based on k-nearest neighbours
in PCA space to assess batch mixing.
}
\examples{
set.seed(1)
bv <- exampleBatchVaria(nGenes = 100)
plotBatchEntropy(bv, assayName = "raw", batchVar = "batch")
bv <- runCorrection(bv, method = "combat", batch = "batch")
plotBatchEntropy(bv, assayName = "raw_combat", batchVar = "batch")

}
